(REF) Convert 'ugprader-base.php.php' to 'Example/UpgraderBase.php'
This makes it considerably easier to read+patch.
