+++ composer.json
@@ -15,6 +15,7 @@
     },
     "autoload": {
         "psr-0": {
+            "CRM_CivixBundle_Resources_Example_": "src/",
             "CRM\\CivixBundle": "src/"
          }
     },
